feat(web): add setCookie util for Next.js server cookies
Add setCookie utility function to parse and set cookie strings using Next.js server-side cookies API.

Handles Set-Cookie string parsing, option extraction, and applies settings for httpOnly, secure, sameSite, path, maxAge, and expires. Throws on invalid Set-Cookie header format.
